azure-servicebus/job-python/requirements.txt
azure-servicebus==7.11.4
